trag in el reel upload

+++ src/MovieApp/MovieApp.Web/Controllers/ReelsUploadController.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.IO;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
@@ -71,10 +71,12 @@ public class ReelsUploadController : Controller
             return RedirectToAction(nameof(Index));
         }
 
-        string tempPath = Path.GetTempFileName();
+        string? tempPath = null;
 
         try
         {
+            tempPath = Path.GetTempFileName();
+
             using (var stream = new FileStream(tempPath, FileMode.Create))
             {
                 await videoFile.CopyToAsync(stream);
@@ -103,7 +105,7 @@ public class ReelsUploadController : Controller
         }
         finally
         {
-            if (System.IO.File.Exists(tempPath))
+            if (tempPath != null && System.IO.File.Exists(tempPath))
             {
                 System.IO.File.Delete(tempPath);
             }
